Optimize getDistinctNamespaces
The current implementation will perform poorly when the desiredset
contains a large number of objects and namespaces, as it does a linear
scan of the string slice for each object. A more optimal approach is to
build a map of namespaces, and just return the map keys. This is much
faster since map operations are O(1).

Signed-off-by: Brad Davidson <brad.davidson@rancher.com>