Ensure map keys are ordered before patch compare
In certain instances, marshaling a struct will produce a different result than marshaling a map with the same JSON object structure. This is because the encoding/json package guarantees ordered keys when marshaling a map. This would produce false change detection when calculating patches for certain objects.

After this change, the runtime.Object will be marshaled to bytes, unmarshaled to a map[string]interface{}, and then this map marshaled to bytes. This ensures that the keys will be ordered and will not produce unexpect patches.