Fix job reconciler diff
For reconcilers the originalObject that was passed is built from the
the applied annotation.  That object will have all string values
truncated to 64 bytes.  The job reconciler would fail because it does
a full object compare and if any desired value was greater that 64 bytes
it would never match because the originalObject's >64 byte value would
be truncated.

To resolve this we now pass the newObject to be created and all the
newObject truncated so that an exact comparsion can be done against
the originalObject.