Add Reconciled condition to TransitioningUnknown
In order to make a better and more consistent experience in Rancher's
UI, some objects will now go into a Reconciled unknown state. In order
for this to showup properly in the UI, this condition needs to be added
to the list of TransitioningUnknown conditions. This change does exactly
that.