leaderelection: exit 0 when exit requested
The leader election code would always terminate fatally, with a
non-zero exit code. Even if the application requested a shutdown.
This was common when paired with wrangler-cli setting up a content
that canceled when SIGTERM was detected. SIGTERM is sent when a node
shuts down to tell the processes to exit, for example. The non-zero
exit code is then recorded as an error by Kubernetes.

Instead, when an exit is requested because the context is set to done
wrangler can cleanly exit. This is a pattern found in the kube scheduler

Signed-off-by: Matt Farina <matt.farina@suse.com>