Code cleanup.
Removes unused member variables
`InformersPackage` `ClientsPackage` & `ListersPackage` from pkg/args.Group.
These variables were previously used to create import statements in our generated code.
Though the refactor in #85 made the variables obsolete.

Removes the generated function DeepCopy<type>OnChange.
This function is unused by our rancher products and does not work as initially intended.
This function's original intent was to be a helper function that creates
a Handler that modifies a copy of the provided object and then
updates the object if the modification produces a change.
This function was modified in 2044a7b2bb07a187dff7a936d6f6a8740d29ad9e
so that it no longer returns a handler.